pyproject.toml
[project]
name = "ocean4dvarnet"
version = "0.0.13"
description = "A computer program whose purpose is to apply deep learning schemes to dynamical systems and ocean remote sensing data."
authors = [
  { name="Ronan Fablet", email="ronan.fablet@imt-atlantique.fr" },
  { name="Quentin Febvre", email="quentin.febvre@ifremer.fr" },
  { name="Pierre Haslee", email="author@example.com" },
  { name="DanieL Zhu", email="daniel.zhu@imt-atlantique.fr" },
  { name="Hugo Georgenthum", email="hugo.georgenthum@imt-atlantique.fr" },
  { name="Braux Emmanuel", email="emmanuel.braux@imt-atlantique.fr" },
]
readme = "README.md"
# Disable license informations when using setuptools as "builder backend"
#  https://github.com/pypa/packaging-problems/issues/870
#  https://github.com/pypa/setuptools/issues/4828
#license = "CECILL-C"
#license-files = ["LICEN[CS]E*"]
requires-python = "~=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "xarray==2025.1.2",
  "netcdf4 (>=1.7.2,<2.0.0)",
  "numpy==1.26.4",
  "pandas",
  "pytorch-lightning==2.3.3",
]  

# conda dependencies
#  - pytorch::pytorch=2.4.1
#   - pytorch::pytorch-cuda=11.8


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pylint.format]
max-line-length = 120